requirements.txt
# ── Core ─────────────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
langchain>=0.3.7
langchain-openai>=0.2.3
openai>=1.40.0

# ── Data & Storage ───────────────────────────────
qdrant-client>=1.11.3
psycopg2-binary>=2.9.9
pandas>=2.2.2

# ── Document Processing ─────────────────────────
pymupdf>=1.24.9

# ── Frontend ─────────────────────────────────────
streamlit>=1.37.1
requests>=2.31.0

# ── Monitoring ───────────────────────────────────
prometheus-client>=0.20.0

# ── Config ───────────────────────────────────────
python-dotenv>=1.0.1
pydantic>=2.7.0

# ── HTTP (transitive, explicit for clarity) ──────
httpx>=0.27.0

# ── Visualisation ────────────────────────────────
plotly>=5.0.0
matplotlib>=3.7.0
